requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
sqlalchemy==2.0.25
psycopg2-binary==2.9.10
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.26.0
jieba==0.42.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
PyJWT==2.8.0
